+++ layers/tests/__init__.py
@@ -7,6 +7,7 @@ from django.test.utils import override_settings
 from django.templatetags.static import static
 from django.contrib.staticfiles import finders
 from django.views.static import serve
+from django.core.urlresolvers import reverse
 from django.conf import settings
 
 BASIC_LAYERS = {'layers': ['basic']}
@@ -158,3 +159,42 @@ class TestCase(BaseTestCase):
     def test_app_bar_static_web(self):
         content = self.get_rendered_static('app_bar.css')
         self.assertEqual(content, 'app bar web')
+
+
+class DecoratorTestCase(TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        cls.request = RequestFactory()
+        cls.request.method = 'GET'
+        cls.request._path = '/'
+        cls.request.get_full_path = lambda: cls.request._path
+        cls.client = Client()
+
+    @override_settings(LAYERS=BASIC_LAYERS)
+    def test_exclude_from_layers_basic(self):
+        url = reverse('normal-view')
+        response = self.client.get(url)
+        result = response.content
+        self.assertEqual(response.status_code, 200)
+        self.failUnless('This is a normal view' in result)
+
+        url = reverse('web-only-view')
+        response = self.client.get(url)
+        result = response.content
+        self.assertEqual(response.status_code, 200)
+        self.failUnless('is not available for your device' in result)
+
+    @override_settings(LAYERS=WEB_LAYERS)
+    def test_exclude_from_layers_web(self):
+        url = reverse('normal-view')
+        response = self.client.get(url)
+        result = response.content
+        self.assertEqual(response.status_code, 200)
+        self.failUnless('This is a normal view' in result)
+
+        url = reverse('web-only-view')
+        response = self.client.get(url)
+        result = response.content
+        self.assertEqual(response.status_code, 200)
+        self.failUnless('This view is only available for web' in result)