+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+#### next
+1. Provide decorator `exclude_from_layers` so a view renders properly even if it can't render for a particular layer.
+
 #### 0.2
 1. Inevitable package rename.